refactor(config): split schemas.rs (2004 lines) into schemas/ directory
Sub-modules: helpers (field builders), controllers (registration),
schema_defs (per-function schema match). Test companion preserved.
